ml-service/requirements.txt
fastapi
uvicorn
transformers
torch
pandas
numpy
scikit-learn
indic-nlp-library
spacy
python-multipart
pydantic
